+++ lib/config.ml
@@ -12,6 +12,13 @@ module Mount = struct
   }
 end
 
+module Secret = struct
+  type t = {
+    value: string;
+    target: string;
+  } [@@deriving sexp]
+end
+
 type t = {
   cwd : string;
   argv : string list;
@@ -20,7 +27,8 @@ type t = {
   env : env;
   mounts : Mount.t list;
   network : string list;
+  mount_secrets : Secret.t list;
 }
 
-let v ~cwd ~argv ~hostname ~user ~env ~mounts ~network =
-  { cwd; argv; hostname; user; env; mounts; network }
+let v ~cwd ~argv ~hostname ~user ~env ~mounts ~network ~mount_secrets =
+  { cwd; argv; hostname; user; env; mounts; network; mount_secrets }