+++ lib/runc_sandbox.ml
@@ -25,6 +25,8 @@ let get_arches () =
     []
   )
 
+let secret_file id = "secret-" ^ string_of_int id
+
 module Json_config = struct
   let mount ?(options=[]) ~ty ~src dst =
     `Assoc [
@@ -99,7 +101,7 @@ module Json_config = struct
     in
     `Assoc fields
 
-  let make {Config.cwd; argv; hostname; user; env; mounts; network} t ~config_dir ~results_dir : Yojson.Safe.t =
+  let make {Config.cwd; argv; hostname; user; env; mounts; network; mount_secrets} t ~config_dir ~results_dir : Yojson.Safe.t =
     let user =
       let { Obuilder_spec.uid; gid } = user in
       `Assoc [
@@ -227,6 +229,17 @@ module Json_config = struct
            ]
          else []
         ) @
+        List.mapi (fun id { Config.Secret.target; _} ->
+          mount target
+            ~ty:"bind"
+            ~src:(config_dir / secret_file id)
+            ~options:[
+              "rbind";
+              "rprivate";
+              "ro";
+            ]
+        ) mount_secrets
+         @
         user_mounts mounts
       );
       "linux", `Assoc [
@@ -253,7 +266,7 @@ module Json_config = struct
         "seccomp", seccomp_policy t;
       ];
     ]
-end  
+end
 
 let next_id = ref 0
 
@@ -271,6 +284,12 @@ let run ~cancelled ?stdin:stdin ~log t config results_dir =
   let json_config = Json_config.make config ~config_dir:tmp ~results_dir t in
   Os.write_file ~path:(tmp / "config.json") (Yojson.Safe.pretty_to_string json_config ^ "\n") >>= fun () ->
   Os.write_file ~path:(tmp / "hosts") "127.0.0.1 localhost builder" >>= fun () ->
+  Lwt_list.fold_left_s
+    (fun id Config.Secret.{value; _} ->
+      Os.write_file ~path:(tmp / secret_file id) value >|= fun () ->
+      id + 1
+    ) 0 config.mount_secrets
+  >>= fun _ ->
   let id = string_of_int !next_id in
   incr next_id;
   Os.with_pipe_from_child @@ fun ~r:out_r ~w:out_w ->