+++ lib_spec/obuilder_spec.ml
@@ -1,4 +1,5 @@
 include Spec
 
 module Cache = Cache
+module Secret = Secret
 module Docker = Docker